More robust dhcpd.c
This code fixes the checksum error when the packet size is odd number. Also it has more robust address assignment to handle the IP address conflict. I can make this address assignment more robust, if you guys want. 

I think since this kind of stuff is not included in the minimum requirement, I think we can add this as the extra feature